Move UserID to `:core:foundation`
Move the `UserID` class from `dev.teogor.ceres.data.datastore.common` to `dev.teogor.ceres.core.foundation.models`.
