+++ core/foundation/api/foundation.api
@@ -279,6 +279,29 @@ public final class dev/teogor/ceres/core/foundation/extensions/MediaPlayerExtKt 
 	public static final fun createMediaPlayer (Landroid/net/Uri;Landroid/view/SurfaceHolder;Landroid/media/AudioAttributes;ILandroidx/compose/runtime/Composer;II)Landroid/media/MediaPlayer;
 }
 
+public final class dev/teogor/ceres/core/foundation/models/UserID {
+	public static final field $stable I
+	public fun <init> ()V
+	public fun <init> (Ljava/lang/String;)V
+	public synthetic fun <init> (Ljava/lang/String;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public final fun component1 ()Ljava/lang/String;
+	public final fun copy (Ljava/lang/String;)Ldev/teogor/ceres/core/foundation/models/UserID;
+	public static synthetic fun copy$default (Ldev/teogor/ceres/core/foundation/models/UserID;Ljava/lang/String;ILjava/lang/Object;)Ldev/teogor/ceres/core/foundation/models/UserID;
+	public fun equals (Ljava/lang/Object;)Z
+	public final fun generate ()Ljava/lang/String;
+	public final fun getValue ()Ljava/lang/String;
+	public fun hashCode ()I
+	public final fun isEmpty ()Z
+	public final fun setValue (Ljava/lang/String;)V
+	public fun toString ()Ljava/lang/String;
+}
+
+public final class dev/teogor/ceres/core/foundation/models/UserIDKt {
+	public static final fun getFormattedValueInRows (Ldev/teogor/ceres/core/foundation/models/UserID;)Ljava/lang/String;
+	public static final fun getPrivacyFormattedValue (Ldev/teogor/ceres/core/foundation/models/UserID;I)Ljava/lang/String;
+	public static synthetic fun getPrivacyFormattedValue$default (Ldev/teogor/ceres/core/foundation/models/UserID;IILjava/lang/Object;)Ljava/lang/String;
+}
+
 public abstract interface class dev/teogor/ceres/core/foundation/ui/platform/ApplicationDetails {
 	public abstract fun getAppName ()Ljava/lang/String;
 	public abstract fun getPrivacyEmail ()Ljava/lang/String;
