+++ data/datastore/api/datastore.api
@@ -29,28 +29,6 @@ public final class dev/teogor/ceres/data/datastore/common/PreferenceDatastoreKt 
 	public static final fun launch (Ldev/teogor/ceres/data/datastore/common/PreferenceDatastore;Lkotlin/jvm/functions/Function1;)Lkotlinx/coroutines/Job;
 }
 
-public final class dev/teogor/ceres/data/datastore/common/UserID {
-	public fun <init> ()V
-	public fun <init> (Ljava/lang/String;)V
-	public synthetic fun <init> (Ljava/lang/String;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public final fun component1 ()Ljava/lang/String;
-	public final fun copy (Ljava/lang/String;)Ldev/teogor/ceres/data/datastore/common/UserID;
-	public static synthetic fun copy$default (Ldev/teogor/ceres/data/datastore/common/UserID;Ljava/lang/String;ILjava/lang/Object;)Ldev/teogor/ceres/data/datastore/common/UserID;
-	public fun equals (Ljava/lang/Object;)Z
-	public final fun generate ()Ljava/lang/String;
-	public final fun getValue ()Ljava/lang/String;
-	public fun hashCode ()I
-	public final fun isEmpty ()Z
-	public final fun setValue (Ljava/lang/String;)V
-	public fun toString ()Ljava/lang/String;
-}
-
-public final class dev/teogor/ceres/data/datastore/common/UserIDKt {
-	public static final fun getFormattedValueInRows (Ldev/teogor/ceres/data/datastore/common/UserID;)Ljava/lang/String;
-	public static final fun getPrivacyFormattedValue (Ldev/teogor/ceres/data/datastore/common/UserID;I)Ljava/lang/String;
-	public static synthetic fun getPrivacyFormattedValue$default (Ldev/teogor/ceres/data/datastore/common/UserID;IILjava/lang/Object;)Ljava/lang/String;
-}
-
 public final class dev/teogor/ceres/data/datastore/defaults/AppTheme : java/lang/Enum {
 	public static final field ClearlyWhite Ldev/teogor/ceres/data/datastore/defaults/AppTheme;
 	public static final field FollowSystem Ldev/teogor/ceres/data/datastore/defaults/AppTheme;
@@ -87,7 +65,7 @@ public final class dev/teogor/ceres/data/datastore/defaults/CeresPreferences : d
 	public final fun getProfileImageFlow ()Lkotlinx/coroutines/flow/Flow;
 	public final fun getThemeSeed ()Ljava/lang/String;
 	public final fun getThemeSeedFlow ()Lkotlinx/coroutines/flow/Flow;
-	public final fun getUserId ()Ldev/teogor/ceres/data/datastore/common/UserID;
+	public final fun getUserId ()Ldev/teogor/ceres/core/foundation/models/UserID;
 	public final fun setAppTheme (Ldev/teogor/ceres/data/datastore/defaults/AppTheme;)V
 	public final fun setCoverImage (Ljava/lang/String;)V
 	public final fun setDisableAndroidTheme (Z)V
@@ -100,7 +78,7 @@ public final class dev/teogor/ceres/data/datastore/defaults/CeresPreferences : d
 	public final fun setOnboardingComplete (Z)V
 	public final fun setProfileImage (Ljava/lang/String;)V
 	public final fun setThemeSeed (Ljava/lang/String;)V
-	public final fun setUserId (Ldev/teogor/ceres/data/datastore/common/UserID;)V
+	public final fun setUserId (Ldev/teogor/ceres/core/foundation/models/UserID;)V
 }
 
 public final class dev/teogor/ceres/data/datastore/defaults/CeresPreferencesKt {
